backend/requirements.txt
Django==4.2.2
djangochannelsrestframework==1.1.0
python-decouple==3.8
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
channels-auth-token-middlewares==1.1.0
channels-redis==4.1.0
djangorestframework-camel-case==1.4.2